[codex] Extract Quick Check evaluation module (#670)
* extract quick check evaluation module

* decouple quick check status from retrieval
